+++ projects/ionic-angular-theme-switch/src/lib/ionic-angular-theme-switch-toggle.module.ts
@@ -1,4 +1,4 @@
-import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
+import { APP_INITIALIZER, CUSTOM_ELEMENTS_SCHEMA, ModuleWithProviders, NgModule } from '@angular/core';
 import { IonicStorageModule } from '@ionic/storage';
 import { IonicAngularThemeSwitchToggleComponent } from './ionic-angular-theme-switch-toggle.component';
 import { IonicAngularThemeSwitchService } from './ionic-angular-theme-switch.service';
@@ -8,10 +8,7 @@ import { IonicAngularThemeSwitchService } from './ionic-angular-theme-switch.ser
     IonicAngularThemeSwitchToggleComponent
   ],
   imports: [
-    IonicStorageModule.forRoot()
-  ],
-  providers: [
-    IonicAngularThemeSwitchService
+    IonicStorageModule
   ],
   exports: [
     IonicAngularThemeSwitchToggleComponent
@@ -20,5 +17,27 @@ import { IonicAngularThemeSwitchService } from './ionic-angular-theme-switch.ser
     CUSTOM_ELEMENTS_SCHEMA
   ]
 })
+
 export class IonicAngularThemeSwitchToggleModule {
+  static forRoot(): ModuleWithProviders {
+    return {
+      ngModule: IonicAngularThemeSwitchToggleModule,
+      providers: [
+        IonicAngularThemeSwitchService,
+        {
+          provide: APP_INITIALIZER,
+          useFactory: bootstrapIonicAngularThemeSwitch,
+          deps: [ IonicAngularThemeSwitchService ],
+          multi: true
+        }
+      ]
+    };
+  }
+}
+
+export function bootstrapIonicAngularThemeSwitch(ionicAngularThemeSwitchService: IonicAngularThemeSwitchService) {
+  // see https://github.com/angular/angular/issues/14485
+  const res = () => {
+  };
+  return res;
 }